Fix RUSTSEC-2020-0071
https://rustsec.org/advisories/RUSTSEC-2020-0071

Although this crate doesn't use any of the affected methods on the
advisory, it transiftively uses the version of the crate with the issue.
This caused cargo deny check to be triggered when adding this to a repo
for testing.

The change in datetime.rs fixes the call to a deprecated method in
chrono. It's also unlikely to be triggered, but fixed for good measure.